Add SpectrumRanker class.
